moved cmds from init file. updated scalars for setters
