+++ pyvesc/examples/motor_example.py
@@ -2,14 +2,14 @@ from pyvesc import VESC
 import time
 
 # serial port that VESC is connected to. Something like "COM3" for windows and as below for linux/mac
-serial_port = '/dev/tty.usbmodem301'
+serial_port = '/dev/serial/by-id/usb-STMicroelectronics_ChibiOS_RT_Virtual_COM_Port_301-if00'
 
 
 # a function to show how to use the class with a with-statement
 def run_motor_using_with():
     with VESC(serial_port=serial_port) as motor:
         print("Firmware: ", motor.get_firmware_version())
-        motor.set_rpm(300)
+        motor.set_duty_cycle(.02)
 
         # run motor and print out rpm for ~2 seconds
         for i in range(30):
